fix about task merge conflicts and apply enhancements
